* refactor: Yogurt 가격 프로퍼티 타입을 int로 변경

* feta: Yogurt 삭제 API 추가

* chore: 사용안하는 파일 삭제, 어노테이션 추가, 함수 인라인화

* refactor: Yogurt 삭제 메소드 수정. 반환 타입을 Void에서 ResponseEntity로 변경
- 삭제하려는 Yogurt ID가 없을 때 NotFound로 반환하기 위해

* test: Yogurt test 케이스 수정 및 추가
- Controller, Repository 분리
- Auth Jwt 생성 케이스 추가

+++ src/main/java/com/zzimkong/h2/yogurt/domain/Yogurt.java
@@ -17,9 +17,9 @@ public class Yogurt {
     private @Id String id;
     private String name;
     private YogurtType type;
-    private double price;
+    private int price;
 
-    public Yogurt(String name, YogurtType type, double price) {
+    public Yogurt(String name, YogurtType type, int price) {
         this.name = name;
         this.type = type;
         this.price = price;
